+++ client/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [0.4.1] - 2021-02-25
 
-- `ConnectUnreachableError` exported.
+- `UnreachableError` exported.
 - Improved documentation.
 
 ## [0.4.0] - 2021-02-24